Bare-bones graft of new lexer/parser into the main code
No new functionality at this point; the new information from the
lexer/parser is stripped out as soon as we receive it, and processing
continues the old way from there. But the graft is successful in that
everything which has always worked will work under the new "paradigm."

Also cleaned up most of the comments in generateWhereClause() (a few
outstanding issues remain; look for the TODOs). Finally, added TODO
annotations to normalizeNCVConstruct() to note how we can integrate the
full power of the new parsing engine into the main code.