Pre-alpha-stage implementation of smart option reduction up to the construction of the list of boolean feasibility indicators of each option in each word list. So far only column name reduction has been implemented.
Yet to be implemented is the actual reduction of the candidate lists
based on the feasibility indicators. But the current state of the
implementation is enough of a milestone to warrant a checkin.

This checkin has NOT been tested or debugged! It is for reference
only and is not likely to run as-is.