Introduce "QQ" (Quick Query) script, deprecating table-specific scripts
QQ with table-specific configurations is the preferred vehicle for
exercising the power of TQL on regular data going forward. (For
metadata and special queries, the supplied specialty scripts such as
FindColumn, Mhead, ... retain pride of place.) This new approach has
been validated up through the loading of a properly-structured config
file. A sample config file to go with the sample DB, qq_sample_db, is
also introduced. We also check in the metadata for this DB and its table
table1 in the "cache" subdirectory.

Revise error messaging internals. Make error stream configurable instead
of hard-coding "/dev/stderr" always and everywhere.

Introduce query type enumeration and API.